remove response to request file
