+++ myproject/settings_client.py
@@ -36,7 +36,7 @@ ALLOWED_HOSTS = []
 INSTALLED_APPS = [
     'django.contrib.admin', 'django.contrib.auth',
     'django.contrib.contenttypes', 'django.contrib.sessions',
-    'django.contrib.messages', 'django.contrib.staticfiles', 'clients', 'myapp'
+    'django.contrib.messages', 'django.contrib.staticfiles', 'clients'
 ]
 
 MIDDLEWARE = [
